requirements.txt
flask
torch
espnet
parallel_wavegan
scipy
